requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
asgiref==3.9.1
attrs==25.3.0
autoflake==2.3.1
black==25.1.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
frozenlist==1.7.0
idna==3.10
isort==6.0.1
multidict==6.6.3
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
requests==2.32.4
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
yarl==1.20.1
